CDH-29426 CDH-29525 distcp to s3a fails due to integer overflow..
Build fixes for newer AWS SDK.

Fix classpath shading for joda-time.

Excluding joda-time from the aws-java-sdk dependency breaks with newer versions
of the aws sdk.

In fact, the previous maven shading config was broken, because the shading
relocation does nothing if that artifact is not a dependency.

This commit adds the joda time classes to the uber/shaded jar, and adds
depenency exclusions for the parent artifacts that depend on this one.

Change-Id: Iea44f29578dacf0e577c3b235628b5823b6ae7f1